(no longer paired programming) Moved filereading methods to utility class so they can be used in CameraActivity and GalleryActivity cleanly
