+++ DangerMoleMobile/app/src/main/java/com/example/dangermolemobile/Utility.kt
@@ -8,6 +8,7 @@ import android.content.pm.PackageManager
 import android.support.v4.content.ContextCompat
 import android.app.Activity
 import android.support.v4.app.ActivityCompat
+import java.io.File
 import java.math.RoundingMode
 import java.text.DecimalFormat
 
@@ -41,6 +42,14 @@ class Utility {
         dec.roundingMode = RoundingMode.CEILING
         number = dec.format(number).toFloat()
         return number.toString()
+    }
+
+    private fun readFileAsTextUsingInputStream(fileName: String)
+            = File(fileName).inputStream().readBytes().toString(Charsets.UTF_8)
+
+    fun populateArrayFromFile(filePath: String):Array<String>{
+        var str = readFileAsTextUsingInputStream(filePath)
+        return str.split("\n").toTypedArray()
     }
 
 }
